digital_gold/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
sqlite3
streamlit
requests
python-multipart